third_party/blink/web_tests/wpt_internal/prerender/restriction_window_open.html

<!DOCTYPE html>
<script src="/common/utils.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

promise_test(async t => {
  // The key used for storing a test result in the server.
  const key = token();

  // Start prerendering a page that attempts to open a window.
  startPrerendering(`resources/window-open.html?key=${key}`);

  // Wait until the prerendered page stores the test result in the server.
  const value = await nextValueFromServer(key);
  assert_equals(
      value, 'failed to open',
      'window.open() should fail in prerendered pages');
}, 'window.open() should fail in same-origin prerendering page');

// TODO(https://crbug.com/1169032): Test window.open() in cross-origin
// prerendering page when we support cross-origin prerendering.

</script>
</body>
